Use dynamic code modules version in the code modules e2e test
